Refactored services components
