+++ src/app/auth/services/auth.service.ts
@@ -1,13 +1,13 @@
 import { Injectable } from '@angular/core';
 import { Router } from '@angular/router';
 import { AngularFireAuth } from '@angular/fire/compat/auth';
-import { TrainingService } from '../training/training.service';
-import { UiService } from '../shared/ui.service';
+import { TrainingService } from '../../training/training.service';
+import { UiService } from '../../shared/ui.service';
 import { Store } from '@ngrx/store';
-import * as fromRoot from '../app.reducer';
-import * as UI from '../shared/ui.actions';
-import * as Auth from './auth.actions';
-import { AuthData } from './auth-data.model';
+import * as fromRoot from '../../app.reducer';
+import * as UI from '../../shared/ui.actions';
+import * as Auth from '../auth.actions';
+import { AuthData } from '../models/auth-data.model';
 
 @Injectable()
 export class AuthService {
@@ -57,5 +57,18 @@ export class AuthService {
 
   logout() {
     this.afAuth.signOut();
+  }
+
+  getUserID(): Promise<string> {
+    console.log('hey');    
+    return new Promise((resolve, reject) => {
+      this.afAuth.authState.subscribe(user => {
+        if (user) {
+          resolve(user.uid);
+        } else {
+          reject('User not found');
+        }
+      });
+    });
   }
 }
