+++ src/app/training/list-exercises/list-exercise.component.ts
@@ -3,7 +3,7 @@ import { Observable } from 'rxjs';
 import * as fromRoot from '../../app.reducer';
 import { Store } from '@ngrx/store';
 import * as fromTraining from '../training.reducer';
-import { Exercise, WeekPlan } from '../exercise.model';
+import { Exercise, WeekPlan } from '../models/exercise.model';
 import { TrainingService } from '../training.service';
 import { MAT_DIALOG_DATA, MatDialog } from '@angular/material/dialog';
 import { EditExerciseComponent } from './edit-exercise.component';
